+++ lib/app/modules/Images/views/images_view.dart
@@ -1,3 +1,4 @@
+import 'package:filemanager/app/modules/Images/components/image_card/image_card.dart';
 import 'package:filemanager/app/modules/Images/components/top_tool_bar/top_tool_bar.dart';
 import 'package:filemanager/constant.dart';
 import 'package:flutter/material.dart';
@@ -27,20 +28,24 @@ class ImagesView extends GetView<ImagesController> {
                 children: [
                   TopToolBar(),
                   Expanded(
-                    child: StaggeredGridView.countBuilder(
-                      padding: EdgeInsets.all(kDefaultPadding),
-                      crossAxisCount: 4,
-                      itemCount: 18,
-                      itemBuilder: (BuildContext context, int index) =>
-                          index == 0
-                              ? Title()
-                              : ImageCard(
-                                  index: index,
-                                ),
-                      staggeredTileBuilder: (int index) =>
-                          new StaggeredTile.count(2, 1),
-                      mainAxisSpacing: 4.0,
-                      crossAxisSpacing: 4.0,
+                    child: Padding(
+                      padding:
+                          const EdgeInsets.symmetric(vertical: kDefaultPadding),
+                      child: StaggeredGridView.countBuilder(
+                        padding: EdgeInsets.all(kDefaultPadding),
+                        crossAxisCount: 4,
+                        itemCount: 18,
+                        itemBuilder: (BuildContext context, int index) =>
+                            index == 0
+                                ? Title()
+                                : ImageCard(
+                                    index: index,
+                                  ),
+                        staggeredTileBuilder: (int index) =>
+                            new StaggeredTile.fit(2),
+                        mainAxisSpacing: kDefaultPadding * 0.6,
+                        crossAxisSpacing: kDefaultPadding * 0.6,
+                      ),
                     ),
                   )
                 ],
@@ -69,27 +74,6 @@ class Title extends StatelessWidget {
         color: const Color(0xff383f4b),
       ),
       textAlign: TextAlign.left,
-    );
-  }
-}
-
-class ImageCard extends StatelessWidget {
-  final index;
-  const ImageCard({
-    Key key,
-    this.index,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return new Container(
-      color: Colors.green,
-      child: new Center(
-        child: new CircleAvatar(
-          backgroundColor: Colors.white,
-          child: new Text('$index'),
-        ),
-      ),
     );
   }
 }
